move HelloWorldActivityTest and MainActivityTest to package activityTests
